[VM] Implement EEI functions. 1. useGas(). 2. getGasLeft(). 3. getCallValue(). 4. getReturnDataSize().

+++ include/vm/hostfunc/ethereum/callStatic.h
@@ -12,7 +12,7 @@ public:
       : EEI(HostEnv, Cost) {}
 
   ErrCode body(VM::EnvironmentManager &EnvMgr,
-               Instance::MemoryInstance &MemInst, uint32_t &Ret, uint32_t Gas,
+               Instance::MemoryInstance &MemInst, uint32_t &Ret, uint64_t Gas,
                uint32_t AddressOffset, uint32_t DataOffset,
                uint32_t DataLength);
 };
